RBAC_GP3/week 5/requirements.txt
# Week 5: FastAPI Backend Dependencies

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database (for future use)
sqlalchemy>=2.0.0

# Utilities
httpx>=0.25.0
